Bump to upstream version 5.16.
I think the only change is that 5.16 incorporates our 'wcsfix-test.patch' fix.
The main motivation for the update is to rebuild the package on OSX, since the
busted pgplot dylib (pgplot-feedstock#1) was resulting in a busted wcslib
dylib as well. Local tests indicate that rebuilding with the new pgplot fixes
things.