Don't duplicate headers
wcslib installs its headers in `$PREFIX/include/wcslib-$MAJOR.$MINOR` and then
symlinks `$PREFIX/include/wcslib` to point at this directory. Conda doesn't
handle symlinks specially when packaging, so this ends up creating two copies
of the headers. In this commit we ditch the symlink and just install to
`$PREFIX/include/wcslib` unconditionally. The wcslib pkg-config file lists
this as the include directory and it's just easier for everyone to deal with,
instead of changing the directory for every minor version increment.